- Added tournaments - basic form example for tournaments (not ajax yet) - removed pyc files from tracking

+++ sunknightsapp/models/discord_server.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class DiscordServer(models.Model):
-    discord_id=models.BigIntegerField(unique=True)
+    discord_id=models.PositiveIntegerField(unique=True)
     name=models.CharField(max_length=30)
 
     def __str__(self):
